Move the upload CSV button to the main nav bar
