+++ frontend/src/metabase/nav/containers/MainNavbar/MainNavbarContainer/MainNavbarView.tsx
@@ -5,7 +5,11 @@ import { t } from "ttag";
 import _ from "underscore";
 
 import ErrorBoundary from "metabase/ErrorBoundary";
-import { useSetting, useUserSetting } from "metabase/common/hooks";
+import {
+  useHasTokenFeature,
+  useSetting,
+  useUserSetting,
+} from "metabase/common/hooks";
 import { useIsAtHomepageDashboard } from "metabase/common/hooks/use-is-at-homepage-dashboard";
 import TippyPopoverWithTrigger from "metabase/components/PopoverWithTrigger/TippyPopoverWithTrigger";
 import { Tree } from "metabase/components/tree";
@@ -18,6 +22,7 @@ import { useSelector } from "metabase/lib/redux";
 import * as Urls from "metabase/lib/urls";
 import { WhatsNewNotification } from "metabase/nav/components/WhatsNewNotification";
 import { getIsEmbedded } from "metabase/selectors/embed";
+import { getSetting } from "metabase/selectors/settings";
 import { getIsWhiteLabeling } from "metabase/selectors/whitelabel";
 import type { IconName, IconProps } from "metabase/ui";
 import type Database from "metabase-lib/v1/metadata/Database";
@@ -35,6 +40,7 @@ import {
   TrashSidebarSection,
 } from "../MainNavbar.styled";
 import { SidebarCollectionLink, SidebarLink } from "../SidebarItems";
+import { DwhUploadCSV } from "../SidebarItems/DwhUploadCSV/DwhUploadCSV";
 import { SidebarOnboardingSection } from "../SidebarItems/SidebarOnboardingSection";
 import { trackOnboardingChecklistOpened } from "../SidebarItems/SidebarOnboardingSection/analytics";
 import type { SelectedItem } from "../types";
@@ -136,6 +142,30 @@ export function MainNavbarView({
   const showOnboardingChecklist =
     isAdmin && showOnboarding && !isEmbedded && !isWhiteLabelled;
 
+  // Instances with DWH enabled already have uploads enabled by default.
+  // It is not possible to turn the uploads off, nor to delete the attached database.
+  const hasAttachedDWHFeature = useHasTokenFeature("attached_dwh");
+
+  const uploadDbId = useSelector(
+    state => getSetting(state, "uploads-settings")?.db_id,
+  );
+
+  /**
+   * the user must have:
+   *   - "write" permissions for the root collection AND
+   *   - "upload" permissions for the attached DWH
+   */
+  const rootCollection = collections.find(
+    c => c.id === "root" || c.id === null,
+  );
+  const canCurateRootCollection = rootCollection?.can_write;
+  const canUploadToDatabase = databases
+    ?.find(db => db.id === uploadDbId)
+    ?.canUpload();
+
+  const canUpload = canCurateRootCollection && canUploadToDatabase;
+  const showUploadCSVButton = hasAttachedDWHFeature && canUpload;
+
   return (
     <ErrorBoundary>
       <SidebarContentRoot>
@@ -160,6 +190,7 @@ export function MainNavbarView({
                 {t`How to use Metabase`}
               </PaddedSidebarLink>
             )}
+            {showUploadCSVButton && <DwhUploadCSV />}
           </SidebarSection>
 
           {bookmarks.length > 0 && (
