+++ frontend/src/metabase/nav/containers/MainNavbar/SidebarItems/DwhUploadCSV/DwhUploadCSV.tsx
@@ -1,0 +1,94 @@
+import { type ChangeEvent, useCallback, useRef, useState } from "react";
+import { t } from "ttag";
+
+import {
+  type CollectionOrTableIdProps,
+  ModelUploadModal,
+} from "metabase/collections/components/ModelUploadModal";
+import type { OnFileUpload } from "metabase/collections/types";
+import { UploadInput } from "metabase/components/upload";
+import { useToggle } from "metabase/hooks/use-toggle";
+import { useDispatch } from "metabase/lib/redux";
+import {
+  type UploadFileProps,
+  uploadFile as uploadFileAction,
+} from "metabase/redux/uploads";
+import { Button, Icon } from "metabase/ui";
+
+import { trackDWHUploadCSVClicked } from "./analytics";
+
+export const DwhUploadCSV = () => {
+  const [
+    isModelUploadModalOpen,
+    { turnOn: openModelUploadModal, turnOff: closeModelUploadModal },
+  ] = useToggle(false);
+
+  const [uploadedFile, setUploadedFile] = useState<File | null>(null);
+
+  const dispatch = useDispatch();
+  const uploadFile = useCallback(
+    ({ file, modelId, collectionId, tableId, uploadMode }: UploadFileProps) =>
+      dispatch(
+        uploadFileAction({ file, modelId, collectionId, tableId, uploadMode }),
+      ),
+    [dispatch],
+  );
+
+  const handleFileUpload = useCallback<OnFileUpload>(
+    (uploadFileArgs: CollectionOrTableIdProps) => {
+      const { collectionId, tableId } = uploadFileArgs;
+      if (uploadedFile && (collectionId || tableId)) {
+        closeModelUploadModal();
+        uploadFile({
+          file: uploadedFile,
+          ...uploadFileArgs,
+        });
+      }
+    },
+    [uploadFile, uploadedFile, closeModelUploadModal],
+  );
+
+  const handleFileInput = (event: ChangeEvent<HTMLInputElement>) => {
+    trackDWHUploadCSVClicked();
+    const file = event.target.files?.[0];
+
+    if (file !== undefined) {
+      setUploadedFile(file);
+      openModelUploadModal();
+
+      // reset the input so that the same file can be uploaded again
+      if (uploadInputRef.current) {
+        uploadInputRef.current.value = "";
+      }
+    }
+  };
+
+  const uploadInputRef = useRef<HTMLInputElement>(null);
+  const triggerUploadInput = () => uploadInputRef.current?.click();
+
+  return (
+    <>
+      <Button
+        color="brand"
+        fullWidth={true}
+        leftIcon={<Icon name="upload" />}
+        onClick={triggerUploadInput}
+        radius="xl"
+        variant="outline"
+      >
+        {t`Upload CSV`}
+      </Button>
+      <UploadInput
+        id="dwh-csv-upload-input"
+        ref={uploadInputRef}
+        onChange={handleFileInput}
+      />
+      <ModelUploadModal
+        collectionId="root"
+        opened={isModelUploadModalOpen}
+        onClose={closeModelUploadModal}
+        onUpload={handleFileUpload}
+      />
+    </>
+  );
+};
