+++ frontend/src/metabase/nav/containers/MainNavbar/SidebarItems/DwhUploadCSV/analytics.ts
@@ -1,0 +1,8 @@
+import { trackSimpleEvent } from "metabase/lib/analytics";
+
+export const trackDWHUploadCSVClicked = () => {
+  trackSimpleEvent({
+    event: "csv_upload_clicked",
+    triggered_from: "left-nav",
+  });
+};
