+++ frontend/src/metabase/nav/containers/MainNavbar/SidebarItems/DwhUploadCSV/index.ts
@@ -1,0 +1,1 @@
+export { DwhUploadCSV } from "./DwhUploadCSV";
