+++ frontend/src/metabase/nav/containers/MainNavbar/SidebarItems/SidebarOnboardingSection/analytics.ts
@@ -1,11 +1,5 @@
 import { trackSimpleEvent } from "metabase/lib/analytics";
 
-export const trackAddDataViaCSV = () => {
-  trackSimpleEvent({
-    event: "data_add_via_csv_clicked",
-  });
-};
-
 export const trackAddDataViaDatabase = () => {
   trackSimpleEvent({
     event: "data_add_via_db_clicked",
